Refactor support code into its own application
