+++ src/thunderbird_accounts/support/__init__.py
@@ -1,0 +1,1 @@
+
