+++ src/thunderbird_accounts/support/apps.py
@@ -1,0 +1,13 @@
+from django.apps import AppConfig
+
+
+class SupportConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'thunderbird_accounts.support'
+    verbose_name = 'Support'
+
+    def ready(self):
+        # Import here so Django finishes app loading before signal registration.
+        from thunderbird_accounts.support.cors import register_cors_handlers
+
+        register_cors_handlers()
