+++ src/thunderbird_accounts/support/tests/__init__.py
@@ -1,0 +1,1 @@
+
