+++ src/thunderbird_accounts/support/tests/test_cors.py
@@ -3,7 +3,7 @@ from django.urls import reverse
 
 
 @override_settings(CORS_ALLOWED_ORIGINS=[], ZENDESK_SUBDOMAIN='example')
-class ZendeskSidebarCorsTestCase(SimpleTestCase):
+class SupportCustomerCorsTestCase(SimpleTestCase):
     def setUp(self):
         self.client = RequestClient()
 
@@ -15,8 +15,8 @@ class ZendeskSidebarCorsTestCase(SimpleTestCase):
             HTTP_ACCESS_CONTROL_REQUEST_HEADERS='authorization, content-type',
         )
 
-    def test_zendesk_origin_is_allowed_for_sidebar_customer_api(self):
-        response = self.options(reverse('api_zendesk_sidebar_customer'), 'https://example.zendesk.com')
+    def test_zendesk_origin_is_allowed_for_support_customer_api(self):
+        response = self.options(reverse('api_support_customer'), 'https://example.zendesk.com')
 
         self.assertEqual(response.headers.get('Access-Control-Allow-Origin'), 'https://example.zendesk.com')
 
@@ -25,7 +25,7 @@ class ZendeskSidebarCorsTestCase(SimpleTestCase):
 
         self.assertNotIn('Access-Control-Allow-Origin', response.headers)
 
-    def test_other_origin_is_not_allowed_for_sidebar_customer_api(self):
-        response = self.options(reverse('api_zendesk_sidebar_customer'), 'https://attacker.example')
+    def test_other_origin_is_not_allowed_for_support_customer_api(self):
+        response = self.options(reverse('api_support_customer'), 'https://attacker.example')
 
         self.assertNotIn('Access-Control-Allow-Origin', response.headers)
