+++ src/thunderbird_accounts/support/zendesk.py
@@ -1,5 +1,6 @@
-import requests
 import mimetypes
+
+import requests
 from django.conf import settings
